+++ src/Api/Service/RevisionGetter.php
@@ -8,7 +8,6 @@ use Mediawiki\DataModel\Revision;
 use RuntimeException;
 use Wikibase\DataModel\ItemContent;
 use Wikibase\DataModel\PropertyContent;
-use Wikibase\DataModel\Entity\Entity;
 use Wikibase\DataModel\Entity\EntityId;
 use Wikibase\DataModel\Entity\Item;
 use Wikibase\DataModel\Entity\Property;
@@ -93,7 +92,7 @@ class RevisionGetter {
 	}
 
 	/**
-	 * @param Entity $entity
+	 * @param Item|Property $entity
 	 *
 	 * @throws RuntimeException
 	 * @return ItemContent|PropertyContent