+++ RELEASENOTES.md
@@ -3,6 +3,7 @@ These are the release notes for the [wikibase-api](README.md).
 ## Version 0.4 (in development)
 
 * Added RevisionsGetter services for getting multiple revisions in as few requests as possible.
+* Stop type hinting against deprecated Entity per https://lists.wikimedia.org/pipermail/wikidata-tech/2014-June/000489.html
 
 ## Version 0.3 (2014-09-15)