+++ RELEASENOTES.md
@@ -1,6 +1,6 @@
 These are the release notes for the [wikibase-api](README.md).
 
-## Version 0.4
+## Version 0.4 (2014-12-14)
 
 * Added RevisionsGetter services for getting multiple revisions in as few requests as possible.
 * Stop type hinting against deprecated Entity per https://lists.wikimedia.org/pipermail/wikidata-tech/2014-June/000489.html